login getting a 500

+++ server/db/migrations/0520181209075758_create.js
@@ -14,7 +14,7 @@ exports.up = function(knex, Promise) {
     table.string('photo', 300);
     table.string('website', 255);
     table.string('description', 1000);
-    table.integer('phone_number');
+    table.bigInteger('phone_number');
     table.string('license_number');
     table.timestamp('created_at').defaultTo(knex.fn.now());
     table.timestamp('updated_at').defaultTo(knex.fn.now());
